clean up renderer namespace, update cmakelists
